server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.35
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
httpx==0.27.2
email-validator==2.2.0   # needed for pydantic.EmailStr
python-multipart==0.0.9  # needed for file uploads (multipart/form-data)
alembic==1.13.1
psycopg2-binary==2.9.10
python-docx==1.1.2
pdfminer.six==20231228